fixed build error
