+++ Assets/Scripts/StrangeEngine/Item.cs
@@ -2,6 +2,7 @@
 // Free to use and modify as you please, Not to be published, distributed, licenced or sold without permission from StrangeDevTeam
 // Requests for the above to be made here: https://www.reddit.com/r/StrangeDev/
 
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -23,42 +24,29 @@ public class Item : ScriptableObject
     public int worth;
     public bool isStackable = false;
     //public Sprite icon;
-    public Dictionary<string, int> itemStats = new Dictionary<string, int>(); // "stats" that come in pairs, a name (string) and a number. for example ("weight", 22)
+    //public Dictionary<string, int> itemStats = new Dictionary<string, int>(); // "stats" that come in pairs, a name (string) and a number. for example ("weight", 22)
 
-    public Item(ItemType pType, int pID, string pName, string pInfo, int pWorth, Dictionary<string,int> pStats)
-    {
-        type = pType;
-        ID = pID;
-        itemName = pName;
-        info = pInfo;
-        worth = pWorth;
-        itemStats = pStats;
-        isStackable = false;
-    }
-    public Item(ItemType pType, int pID, string pName, string pInfo, int pWorth, Dictionary<string, int> pStats, bool pIsStackable)
-    {
-        type = pType;
-        ID = pID;
-        itemName = pName;
-        info = pInfo;
-        worth = pWorth;
-        itemStats = pStats;
-        isStackable = pIsStackable;
-    }
-
-}
-
-[CreateAssetMenu(fileName = "Item", menuName = "StrangeEngine/Weapon", order = 3)]
-public class Weapon : Item
-{
-    public int Damage;
-    public GameObject WeaponModelPrefab;
-
-
-    public Weapon(int pID, string pName, string pInfo, int pWorth, Dictionary<string, int> pStatsfloat, int pDamage, GameObject pWeaponPrefab) : base(ItemType.Equipment,pID, pName, pInfo,pWorth,pStatsfloat)
-    {
-        Damage = pDamage;
-        WeaponModelPrefab = pWeaponPrefab;
-    }
+    //public Item(ItemType pType, int pID, string pName, string pInfo, int pWorth, Dictionary<string,int> pStats)
+    //{
+    //    type = pType;
+    //    ID = pID;
+    //    itemName = pName;
+    //    info = pInfo;
+    //    worth = pWorth;
+    //    //itemStats = pStats;
+    //}
+    //public Item(ItemType pType, int pID, string pName, string pInfo, int pWorth, Dictionary<string, int> pStats, bool pIsStackable)
+    //{
+    //    type = pType;
+    //    ID = pID;
+    //    itemName = pName;
+    //    info = pInfo;
+    //    worth = pWorth;
+    //    //itemStats = pStats;
+    //    isStackable = pIsStackable;
+    //}
 
 }
+
+
+
