+++ Assets/Scripts/StrangeEngine/Weapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d3fd377703b62bb4fa914f645bdff397
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
